Add files via upload
A predictive model for the association of tennis professional players with the use of various features such as player statistics, past performance, and physical attributes to predict the likelihood of a player's success in future matches and tournaments. This machine learning used the linear regression analysis to make its predictions. The model  could also take into account external factors such as injuries, changes in coaching staff, and other relevant information if needed or for further features additions. The ultimate goal of this model would be to provide accurate and actionable insights for coaches, players, and other stakeholders in the tennis industry.